Register events and locations as Alchemy module
The event and location moduls are custom models from the dummy app we used to test the resources controller from Alchemy. By registering them as modules we can also use them when booting the dummy app locally in development.
