+++ spec/dummy/app/controllers/admin/locations_controller.rb
@@ -1,0 +1,3 @@
+class Admin::LocationsController < Alchemy::Admin::ResourcesController
+
+end
