+++ spec/dummy/app/models/location.rb
@@ -1,3 +1,4 @@
 class Location < ActiveRecord::Base
+  include Alchemy::Taggable
   has_many :events
 end
