+++ spec/dummy/config/locales/en.yml
@@ -22,3 +22,14 @@ en:
       essence_richtext: This content type (Essence) represents richtext
       essence_select: This content type (Essence) represents values the editor can choose from
       essence_text: This content type (Essence) represents a simple line of text
+
+  activerecord:
+    models:
+      event:
+        one: Event
+        other: Events
+    attributes:
+      event:
+        tag_names: Tags
+      location:
+        tag_names: Tags
