+++ spec/dummy/config/routes.rb
@@ -6,6 +6,7 @@ Rails.application.routes.draw do
 
   namespace :admin do
     resources :events
+    resources :locations
   end
 
   mount Alchemy::Engine => "/"
