+++ spec/libraries/resource_spec.rb
@@ -345,7 +345,7 @@ module Alchemy
 
       describe '#model_association_names' do
         it 'returns an array of association names' do
-          expect(resource.model_association_names).to eq [:location]
+          expect(resource.model_association_names).to include(:location)
         end
       end
